requirements.txt
google-genai
chromadb
pydantic
python-dotenv
sentence-transformers
google-search-results
fastmcp
